backend/requirements.txt
yt_dlp
flask-cors
Flask
colorama
rich
pyqt6
PyQt6-WebEngine
eyed3
python-dotenv